backend/pyproject.toml
[project]
name = "tender-backend"
version = "0.1.0"
description = "Business API for the tender authoring system."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "pydantic-settings>=2.4,<3.0",
  "psycopg[binary,pool]>=3.2,<4.0",
  "python-multipart>=0.0.9,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "alembic>=1.13,<2.0",
  "structlog>=24.1,<25.0",
  "celery[redis]>=5.4,<6.0",
  "docxtpl>=0.18,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1.0",
  "pytest>=8.3,<9.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["tender_backend"]

[tool.pytest.ini_options]
testpaths = ["tests"]
